+++ cocos/platform/mac/CCApplication-mac.h
@@ -52,7 +52,7 @@ public:
     @brief	Callback by Director for limit FPS.
     @param interval The time, which expressed in second in second, between current frame and next.
     */
-    virtual void setAnimationInterval(double interval);
+    virtual void setAnimationInterval(float interval);
         
     /**
     @brief	Get status bar rectangle in GLView window.
